refactor: user domain 엔티티 패키지 생성 및 이동 #5
