+++ src/main/java/org/songeun/petdongne_server/review/domain/entity/ResidenceReview.java
@@ -9,7 +9,7 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.NoArgsConstructor;
 import org.songeun.petdongne_server.residentialComplex.domain.entity.ResidentialComplex;
-import org.songeun.petdongne_server.user.domain.User;
+import org.songeun.petdongne_server.user.domain.entity.User;
 import org.songeun.petdongne_server.global.common.BaseEntity;
 
 @Entity
