+++ src/main/java/org/songeun/petdongne_server/user/infrastructure/UserRepository.java
@@ -1,6 +1,6 @@
 package org.songeun.petdongne_server.user.infrastructure;
 
-import org.songeun.petdongne_server.user.domain.User;
+import org.songeun.petdongne_server.user.domain.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
